[tui] Fix child insertion offsets & colocate tests
- Fix logic in `update_insertion_pos_for_next_box` to correctly preserve
  parent origin offsets instead of zeroing out preserved axes when the
  parent origins are non-zero.
- Update outdated `box_cursor_pos` rustdocs with proper intra-doc links
  to `FlexBox::insertion_pos_for_next_box`.
- Colocate unit tests natively into `surface.rs` to clean up the
  `layout/` directory and conform with testing taxonomies.
- Update v0.8.0 release meta-task to mark PR #452 as complete.

Task: pr-452-fix.md, prepare-v0.8.0-meta-task.md
