+++ .vscode/settings.json
@@ -48,6 +48,7 @@
     "btop",
     "callsite",
     "CBOR",
+    "cecton",
     "cfmakeraw",
     "cfssl",
     "chrono",
@@ -211,6 +212,7 @@
     "lifecycles",
     "linkification",
     "linkify",
+    "linkifying",
     "litint",
     "Loby",
     "localpaintedeffect",
