+++ task/prepare-v0.8.0-meta-task.md
@@ -10,7 +10,7 @@
   - [x] [improve-immature-vt100-shim.md](improve-immature-vt100-shim.md)
   - [x] [pr-448-fix.md](done/pr-448-fix.md) - https://github.com/r3bl-org/r3bl-open-core/pull/448
 - [ ] RRT API (Pending PRs & Issues)
-  - [ ] https://github.com/r3bl-org/r3bl-open-core/pull/452
+  - [x] [pr-452-fix.md](pr-452-fix.md) - https://github.com/r3bl-org/r3bl-open-core/pull/452
   - [ ] https://github.com/r3bl-org/r3bl-open-core/issues/451
 - [ ] Core Types Refactor
   - [ ] [upgrade-range-for-rust_1_96_0.md](upgrade-range-for-rust_1_96_0.md)
