+++ tui/src/tui/layout/surface.rs
@@ -2,9 +2,8 @@
 
 use super::{FlexBox, FlexBoxProps, LayoutDirection, LayoutError, LayoutErrorType,
             LayoutManagement, PerformPositioningAndSizing, SurfaceProps};
-use crate::{throws, unwrap_or_err};
 use crate::{CommonResult, InlineVec, Pos, RenderPipeline, ReqSizePc, Size, TuiStyle,
-            TuiStylesheet, height, width};
+            TuiStylesheet, height, throws, unwrap_or_err, width};
 
 /// Represents a rectangular area of the terminal screen, and not necessarily the full
 /// terminal screen.
@@ -144,12 +143,38 @@ impl PerformPositioningAndSizing for Surface {
 
     fn no_boxes_added(&self) -> bool { self.stack_of_boxes.is_empty() }
 
+    /// Calculates the coordinates where the next sibling box should be drawn, after we've
+    /// allocated space for the current one.
+    ///
+    /// Imagine you have a [`LayoutDirection::Horizontal`] container. When you add a child
+    /// box to it, the next child should be placed directly to its right. This means the
+    /// [row] (Y-coordinate) should remain exactly the same as the parent container, while
+    /// the [column] (X-coordinate) advances.
+    ///
+    /// ```text
+    /// ╭───────────────────────────────────────╮  <-- Parent Container at col: 0, row: 5
+    /// │                                       │
+    /// │ ╭─────────────╮ ╭─────────────╮       │
+    /// │ │ Child 1     │ │ Child 2     │       │
+    /// │ │ (0, 5)      │ │ (50, 5)     │       │
+    /// │ ╰─────────────╯ ╰─────────────╯       │
+    /// │                 ▲                     │
+    /// │                 │                     │
+    /// │                 ╰─ Next insertion pos preserves the parent's row (5).
+    /// │                    Only the col advances (0 + 50).
+    /// ╰───────────────────────────────────────╯
+    /// ```
+    ///
+    /// This updates the [`FlexBox::insertion_pos_for_next_box`] of the current
+    /// [`FlexBox`].
+    ///
+    /// Returns the [Pos] where the next [`FlexBox`] can be added to the stack of boxes.
+    ///
     /// Must be called *before* the new [`FlexBox`] is added to the stack of boxes
     /// otherwise [`LayoutErrorType::ErrorCalculatingNextBoxPos`] error is returned.
     ///
-    /// This updates the `box_cursor_pos` of the current [`FlexBox`].
-    ///
-    /// Returns the [Pos] where the next [`FlexBox`] can be added to the stack of boxes.
+    /// [column]: crate::ColIndex
+    /// [row]: crate::RowIndex
     fn update_insertion_pos_for_next_box(
         &mut self,
         allocated_size: Size,
@@ -162,15 +187,21 @@ impl PerformPositioningAndSizing for Surface {
           LayoutErrorType::ErrorCalculatingNextBoxPos
         };
 
-        let new_pos: Pos = current_insertion_pos + allocated_size;
+        let new_pos = current_insertion_pos + allocated_size;
 
-        // Adjust `new_pos` using Direction.
-        let new_pos: Pos = match current_box.dir {
-            LayoutDirection::Vertical => new_pos * (width(0) + height(1)),
-            LayoutDirection::Horizontal => new_pos * (width(1) + height(0)),
+        // Adjust `new_pos` using Direction. Preserve the non-directional component:
+        // - for vertical keep the column,
+        // - for horizontal keep the row.
+        let new_pos = match current_box.dir {
+            LayoutDirection::Vertical => {
+                new_pos.row_index + current_insertion_pos.col_index
+            }
+            LayoutDirection::Horizontal => {
+                current_insertion_pos.row_index + new_pos.col_index
+            }
         };
 
-        // Update the box_cursor_pos of the current layout.
+        // Update the insertion_pos_for_next_box of the current layout.
         current_box.insertion_pos_for_next_box = new_pos.into();
 
         Ok(new_pos)
@@ -349,5 +380,522 @@ fn cascade_styles(
         None
     } else {
         TuiStylesheet::compute(&Some(style_vec))
+    }
+}
+
+#[cfg(test)]
+mod test_surface_2_col_complex {
+    use crate::{CommonResult, FlexBoxId, FlexBoxProps, LayoutDirection,
+                LayoutManagement, Surface, SurfaceProps, TuiStylesheet, assert_eq2,
+                box_end, box_start, ch, col, console_log, get_tui_styles, height,
+                new_style, req_size_pc, row, throws, throws_with_return, tui_color,
+                tui_stylesheet, width};
+
+    #[test]
+    fn test_surface_2_col_complex() -> CommonResult<()> {
+        throws!({
+            let mut surface = Surface {
+                stylesheet: dsl_stylesheet()?,
+                ..Default::default()
+            };
+
+            surface.surface_start(SurfaceProps {
+                pos: col(0) + row(0),
+                size: width(500) + height(500),
+            })?;
+
+            create_main_container(&mut surface)?;
+
+            surface.surface_end()?;
+
+            println!("{:?}", surface.render_pipeline);
+        });
+    }
+
+    /// Main container "container".
+    fn create_main_container(surface: &mut Surface) -> CommonResult<()> {
+        fn make_container_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let layout_item = surface.stack_of_boxes.first().unwrap();
+                assert_eq2!(layout_item.id, FlexBoxId::from(0));
+                assert_eq2!(layout_item.dir, LayoutDirection::Horizontal);
+
+                assert_eq2!(layout_item.origin_pos, col(0) + row(0));
+                assert_eq2!(layout_item.bounds_size, width(500) + height(500));
+
+                assert_eq2!(layout_item.style_adjusted_origin_pos, col(1) + row(1)); // due to `padding: 1`
+                assert_eq2!(
+                    layout_item.style_adjusted_bounds_size,
+                    width(498) + height(498)
+                ); // due to `padding: 1`
+
+                assert_eq2!(
+                    layout_item.requested_size_percent,
+                    req_size_pc!(width:100, height:100)
+                );
+
+                assert_eq2!(
+                    layout_item.insertion_pos_for_next_box,
+                    Some(col(0) + row(0))
+                );
+
+                assert!(layout_item.get_computed_style().is_some());
+                assert_eq2!(
+                    layout_item.get_computed_style().unwrap().padding,
+                    Some(ch(1))
+                );
+            });
+        }
+
+        throws!({
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(0),
+                dir: LayoutDirection::Horizontal,
+                requested_size_percent: req_size_pc!(width:100, height:100),
+                maybe_styles: get_tui_styles! { @from: surface.stylesheet, [0] },
+            })?;
+
+            make_container_assertions(surface)?;
+
+            create_left_col(surface)?;
+            create_right_col(surface)?;
+
+            surface.box_end()?;
+        });
+    }
+
+    /// Left column 1.
+    fn create_left_col(surface: &mut Surface) -> CommonResult<()> {
+        fn make_left_col_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let layout_item = surface.stack_of_boxes.last().unwrap();
+                assert_eq2!(layout_item.id, FlexBoxId::from(1));
+                assert_eq2!(layout_item.dir, LayoutDirection::Vertical);
+
+                assert_eq2!(layout_item.origin_pos, col(0) + row(0));
+                assert_eq2!(layout_item.bounds_size, width(250) + height(500));
+
+                console_log!(layout_item);
+
+                assert_eq2!(layout_item.style_adjusted_origin_pos, col(3) + row(3)); // Take padding into account.
+                assert_eq2!(
+                    layout_item.style_adjusted_bounds_size,
+                    width(244) + height(494)
+                ); // Take padding into account.
+
+                assert_eq2!(
+                    layout_item.requested_size_percent,
+                    req_size_pc!(width:50, height:100)
+                );
+                assert_eq2!(layout_item.insertion_pos_for_next_box, None);
+
+                assert_ne!(
+                    layout_item.get_computed_style(),
+                    TuiStylesheet::compute(&surface.stylesheet.find_styles_by_ids(&[1]))
+                );
+            });
+        }
+
+        throws!({
+            // With macro.
+            box_start! {
+              in:                     surface,
+              id:                     FlexBoxId::from(1),
+              dir:                    LayoutDirection::Vertical,
+              requested_size_percent: req_size_pc!(width:50, height:100),
+              styles:                 [1]
+            }
+            make_left_col_assertions(surface)?;
+            box_end!(in: surface);
+        });
+    }
+
+    /// Right column 2.
+    fn create_right_col(surface: &mut Surface) -> CommonResult<()> {
+        fn make_right_col_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let current_box = surface.stack_of_boxes.last().unwrap();
+                assert_eq2!(current_box.id, FlexBoxId::from(2));
+                assert_eq2!(current_box.dir, LayoutDirection::Vertical);
+
+                assert_eq2!(current_box.origin_pos, col(250) + row(0));
+                assert_eq2!(current_box.bounds_size, width(250) + height(500));
+
+                assert_eq2!(current_box.style_adjusted_origin_pos, col(254) + row(4)); // Take padding into account.
+                assert_eq2!(
+                    current_box.style_adjusted_bounds_size,
+                    width(242) + height(492)
+                ); // Take padding into account.
+
+                assert_eq2!(
+                    current_box.requested_size_percent,
+                    req_size_pc!(width:50, height:100)
+                );
+                assert_eq2!(current_box.insertion_pos_for_next_box, None);
+
+                assert_ne!(
+                    current_box.get_computed_style(),
+                    TuiStylesheet::compute(&surface.stylesheet.find_styles_by_ids(&[2]))
+                );
+            });
+        }
+
+        throws!({
+            // No macro.
+            surface.box_start(FlexBoxProps {
+                maybe_styles: get_tui_styles! { @from: surface.stylesheet, [2] },
+                id: FlexBoxId::from(2),
+                dir: LayoutDirection::Vertical,
+                requested_size_percent: req_size_pc!(width:50, height:100),
+            })?;
+            make_right_col_assertions(surface)?;
+            surface.box_end()?;
+        });
+    }
+
+    /// Creates a stylesheet containing styles using DSL.
+    fn dsl_stylesheet() -> CommonResult<TuiStylesheet> {
+        throws_with_return!({
+            tui_stylesheet! {
+                new_style!(
+                    id: {0}
+                    padding: {1}
+                ),
+                new_style!(
+                    id: {1}
+                    dim bold
+                    padding: {2}
+                    color_fg: {tui_color!(255, 255, 0)} /* Yellow. */
+                    color_bg: {tui_color!(128, 128, 128)} /* Grey. */
+                ),
+                new_style!(
+                    id: {2}
+                    underline strikethrough
+                    padding: {3}
+                    color_fg: {tui_color!(black)} /* Black. */
+                    color_bg: {tui_color!(white)} /* White. */
+                )
+            }
+        })
+    }
+}
+
+#[cfg(test)]
+mod test_surface_2_col_simple {
+    use crate::{CommonResult, FlexBoxId, FlexBoxProps, LayoutDirection,
+                LayoutManagement, Surface, SurfaceProps, TuiStylesheet, assert_eq2,
+                box_end, box_start, col, get_tui_styles, height, new_style, req_size_pc,
+                row, throws, throws_with_return, tui_color, tui_stylesheet, width};
+
+    #[test]
+    fn test_surface_2_col_simple() -> CommonResult<()> {
+        throws!({
+            let mut surface = Surface {
+                stylesheet: dsl_stylesheet()?,
+                ..Default::default()
+            };
+
+            surface.surface_start(SurfaceProps {
+                pos: col(0) + row(0),
+                size: width(500) + height(500),
+            })?;
+
+            create_main_container(&mut surface)?;
+
+            surface.surface_end()?;
+
+            println!("{:?}", surface.render_pipeline);
+        });
+    }
+
+    /// Main container 0.
+    fn create_main_container(surface: &mut Surface) -> CommonResult<()> {
+        fn make_container_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let layout_item = surface.stack_of_boxes.first().unwrap();
+                assert_eq2!(layout_item.id, FlexBoxId::from(0));
+                assert_eq2!(layout_item.dir, LayoutDirection::Horizontal);
+                assert_eq2!(layout_item.origin_pos, col(0) + row(0));
+                assert_eq2!(layout_item.bounds_size, width(500) + height(500)); // due to `padding: 1`
+                assert_eq2!(
+                    layout_item.requested_size_percent,
+                    req_size_pc!(width:100, height:100)
+                );
+                assert_eq2!(
+                    layout_item.insertion_pos_for_next_box,
+                    Some(col(0) + row(0))
+                );
+                assert_eq2!(layout_item.get_computed_style(), None);
+            });
+        }
+
+        throws!({
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(0),
+                dir: LayoutDirection::Horizontal,
+                requested_size_percent: req_size_pc!(width:100, height:100),
+                maybe_styles: None,
+            })?;
+
+            make_container_assertions(surface)?;
+
+            create_left_col(surface)?;
+            create_right_col(surface)?;
+
+            surface.box_end()?;
+        });
+    }
+
+    /// Left column 1.
+    fn create_left_col(surface: &mut Surface) -> CommonResult<()> {
+        fn make_left_col_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let layout_item = surface.stack_of_boxes.last().unwrap();
+                assert_eq2!(layout_item.id, FlexBoxId::from(1));
+                assert_eq2!(layout_item.dir, LayoutDirection::Vertical);
+
+                assert_eq2!(layout_item.origin_pos, col(0) + row(0));
+                assert_eq2!(layout_item.bounds_size, width(250) + height(500));
+
+                assert_eq2!(layout_item.style_adjusted_origin_pos, col(2) + row(2)); // Take padding into account.
+                assert_eq2!(
+                    layout_item.style_adjusted_bounds_size,
+                    width(246) + height(496)
+                ); // Take padding into account.
+
+                assert_eq2!(
+                    layout_item.requested_size_percent,
+                    req_size_pc!(width:50, height:100)
+                );
+                assert_eq2!(layout_item.insertion_pos_for_next_box, None);
+                assert_eq2!(
+                    layout_item.get_computed_style(),
+                    TuiStylesheet::compute(&surface.stylesheet.find_styles_by_ids(&[1]))
+                );
+            });
+        }
+
+        throws! {{
+            // With macro.
+            box_start! {
+                in:                     surface,
+                id:                     FlexBoxId::from(1),
+                dir:                    LayoutDirection::Vertical,
+                requested_size_percent: req_size_pc!(width:50, height:100),
+                styles:                 [1]
+            }
+            make_left_col_assertions(surface)?;
+            box_end!(in: surface);
+
+        }}
+    }
+
+    /// Right column 2.
+    fn create_right_col(surface: &mut Surface) -> CommonResult<()> {
+        fn make_right_col_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let current_box = surface.stack_of_boxes.last().unwrap();
+                assert_eq2!(current_box.id, FlexBoxId::from(2));
+                assert_eq2!(current_box.dir, LayoutDirection::Vertical);
+
+                assert_eq2!(current_box.origin_pos, col(250) + row(0));
+                assert_eq2!(current_box.bounds_size, width(250) + height(500));
+
+                assert_eq2!(current_box.style_adjusted_origin_pos, col(253) + row(3)); // Take padding into account.
+                assert_eq2!(
+                    current_box.style_adjusted_bounds_size,
+                    width(244) + height(494)
+                ); // Take padding into account.
+
+                assert_eq2!(
+                    current_box.requested_size_percent,
+                    req_size_pc!(width: 50, height: 100)
+                );
+                assert_eq2!(current_box.insertion_pos_for_next_box, None);
+                assert_eq2!(
+                    current_box.get_computed_style(),
+                    TuiStylesheet::compute(&surface.stylesheet.find_styles_by_ids(&[2]))
+                );
+            });
+        }
+
+        throws!({
+            // No macro.
+            surface.box_start(FlexBoxProps {
+                maybe_styles: get_tui_styles! { @from: surface.stylesheet, [2] },
+                id: FlexBoxId::from(2),
+                dir: LayoutDirection::Vertical,
+                requested_size_percent: req_size_pc!(width:50, height:100),
+            })?;
+            make_right_col_assertions(surface)?;
+            surface.box_end()?;
+        });
+    }
+
+    /// Creates a stylesheet containing styles using DSL.
+    fn dsl_stylesheet() -> CommonResult<TuiStylesheet> {
+        throws_with_return!({
+            tui_stylesheet! {
+                new_style!(
+                    id: {1}
+                    dim bold
+                    padding: {2}
+                    color_fg: {tui_color!(255, 255, 0)} /* Yellow. */
+                    color_bg: {tui_color!(128, 128, 128)} /* Grey. */
+                ),
+                new_style!(
+                    id: {2}
+                    underline strikethrough
+                    padding: {3}
+                    color_fg: {tui_color!(0, 0, 0)} /* Black. */
+                    color_bg: {tui_color!(255, 255, 255)} /* White. */
+                )
+            }
+        })
+    }
+}
+
+#[cfg(test)]
+mod test_surface_offset_origin {
+    use crate::{assert_eq2, col, height, req_size_pc, row, throws, width,
+                CommonResult, FlexBoxId, FlexBoxProps, LayoutDirection,
+                LayoutManagement, Surface, SurfaceProps};
+
+    #[test]
+    fn test_surface_horizontal_non_zero_origin() -> CommonResult<()> {
+        fn make_container_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let container = surface.stack_of_boxes.first().unwrap();
+                assert_eq2!(container.origin_pos, col(0) + row(5));
+                assert_eq2!(container.dir, LayoutDirection::Horizontal);
+            });
+        }
+
+        fn make_child1_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let child = surface.stack_of_boxes.last().unwrap();
+                // First child starts at the container's insertion start = (0, 5).
+                assert_eq2!(child.origin_pos, col(0) + row(5));
+            });
+        }
+
+        fn make_child2_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let child = surface.stack_of_boxes.last().unwrap();
+                // Second child row must equal the container's row (5), NOT 0.
+                // Before the fix this was `col(50) + row(0)`.
+                assert_eq2!(child.origin_pos, col(50) + row(5));
+            });
+        }
+
+        throws!({
+            let mut surface = Surface::default();
+
+            // Surface starts at row 5 – the existing tests all start at (0, 0).
+            surface.surface_start(SurfaceProps {
+                pos: col(0) + row(5),
+                size: width(100) + height(40),
+            })?;
+
+            // Horizontal container – no padding, fills the surface.
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(0),
+                dir: LayoutDirection::Horizontal,
+                requested_size_percent: req_size_pc!(width: 100, height: 100),
+                maybe_styles: None,
+            })?;
+            make_container_assertions(&surface)?;
+
+            // First child: 50 % width.
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(1),
+                dir: LayoutDirection::Vertical,
+                requested_size_percent: req_size_pc!(width: 50, height: 100),
+                maybe_styles: None,
+            })?;
+            make_child1_assertions(&surface)?;
+            surface.box_end()?;
+
+            // Second child: 50 % width.
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(2),
+                dir: LayoutDirection::Vertical,
+                requested_size_percent: req_size_pc!(width: 50, height: 100),
+                maybe_styles: None,
+            })?;
+            make_child2_assertions(&surface)?;
+            surface.box_end()?;
+
+            surface.box_end()?;
+            surface.surface_end()?;
+        });
+    }
+
+    #[test]
+    fn test_surface_vertical_non_zero_origin() -> CommonResult<()> {
+        fn make_container_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let container = surface.stack_of_boxes.first().unwrap();
+                assert_eq2!(container.origin_pos, col(5) + row(0));
+                assert_eq2!(container.dir, LayoutDirection::Vertical);
+            });
+        }
+
+        fn make_child1_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let child = surface.stack_of_boxes.last().unwrap();
+                assert_eq2!(child.origin_pos, col(5) + row(0));
+            });
+        }
+
+        fn make_child2_assertions(surface: &Surface) -> CommonResult<()> {
+            throws!({
+                let child = surface.stack_of_boxes.last().unwrap();
+                // Second child column must equal the container's column (5), NOT 0.
+                // Before the fix for Vertical this would have been `col(0) + row(20)`.
+                assert_eq2!(child.origin_pos, col(5) + row(20));
+            });
+        }
+
+        throws!({
+            let mut surface = Surface::default();
+
+            surface.surface_start(SurfaceProps {
+                pos: col(5) + row(0),
+                size: width(100) + height(40),
+            })?;
+
+            // Vertical container - no padding, fills the surface.
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(0),
+                dir: LayoutDirection::Vertical,
+                requested_size_percent: req_size_pc!(width: 100, height: 100),
+                maybe_styles: None,
+            })?;
+            make_container_assertions(&surface)?;
+
+            // First child: 50 % height.
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(1),
+                dir: LayoutDirection::Horizontal,
+                requested_size_percent: req_size_pc!(width: 100, height: 50),
+                maybe_styles: None,
+            })?;
+            make_child1_assertions(&surface)?;
+            surface.box_end()?;
+
+            // Second child: 50 % height.
+            surface.box_start(FlexBoxProps {
+                id: FlexBoxId::from(2),
+                dir: LayoutDirection::Horizontal,
+                requested_size_percent: req_size_pc!(width: 100, height: 50),
+                maybe_styles: None,
+            })?;
+            make_child2_assertions(&surface)?;
+            surface.box_end()?;
+
+            surface.box_end()?;
+            surface.surface_end()?;
+        });
     }
 }
